feat: update test sort.test tsx
