+++ project/package.json
@@ -69,6 +69,7 @@
   "devDependencies": {
     "@jedmao/redux-mock-store": "^3.0.5",
     "@redux-devtools/extension": "^3.2.3",
+    "@testing-library/dom": "9.2.0",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^14.4.3",
